requirements.txt
# Core Libraries
python-dotenv
numpy
pandas
scipy
pytz
redis>=4.0.0,<5.0.0 # Specify async compatible version
aiohttp
websockets
pydantic
# Or torch-cpu depending on setup
# NOTE: For GPU support, install via ./install_xgboost_gpu.sh script

# Broker/API Clients (If installed via pip, otherwise handled by local code)
alpaca-trade-api # Uncomment if using pip install

# ML/Ops Tools (Optional, used by removed/offline components)
# optuna
# shap
# alibi-detect
pybreaker # Circuit breaker library
asyncio-throttle # Rate limiting
prometheus-client # Metrics

# Scheduling
APScheduler>=3.0.0,<4.0.0 # Added Scheduler

# Development/Testing
pytest
pytest-asyncio
python-json-logger # Added for logging
joblib # Added for ML model saving/loading
